requirements.txt
streamlit
pandas
numpy
scikit-learn
yfinance
pyrebase4
setuptools
plotly
httpx
python-dotenv